requirements.txt
cycler==0.10.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
Pillow==8.4.0
pyparsing==3.0.0
python-dateutil==2.8.2
six==1.16.0
torch==1.10.0
torchvision==0.11.1
typing-extensions==3.10.0.2
